+++ utils/check-area.sh
@@ -38,7 +38,7 @@ categorize pmem      -e "^src/libpmem/"     -e "^src/include/libpmem.h"
 categorize pmem2     -e "^src/libpmem2/"    -e "^src/include/libpmem2.h" -e "^src/include/libpmem2/"
 categorize rpmem     -e "^src/librpmem/"    -e "^src/include/librpmem.h" -e "^src/tools/rpmemd/" -e "^src/rpmem_common/"
 categorize log       -e "^src/libpmemlog/"  -e "^src/include/libpmemlog.h"
-categorize blk       -e "^src/libpmemblk/"  -e "^src/include/libpmemblk.h"
+categorize blk       -e "^src/libpmemblk/"  -e "^src/include/libpmemblk.h" -e "^src/include/libpmemblk/"
 categorize obj       -e "^src/libpmemobj/"  -e "^src/include/libpmemobj.h" -e "^src/include/libpmemobj/"
 categorize pool      -e "^src/libpmempool/" -e "^src/include/libpmempool.h" -e "^src/tools/pmempool/"
 categorize set       -e "^src/libpmemset/"  -e "^src/include/libpmemset.h"
